feat: 체험단 MY > 활동 페이지 > 체험단이 찜한 스토어의 정보 가져오기 API 구현 완료

+++ src/main/java/com/nuguna/freview/customer/dto/request/CustomerMyZzimedStoresRetrieveRequestDTO.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 public class CustomerMyZzimedStoresRetrieveRequestDTO {
 
   @Min(1)
-  private Long targetPage;
+  private Integer targetPage;
 }
